requirements.txt
asttokens==2.2.1
backcall==0.2.0
comm==0.1.3
contourpy==1.0.7
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
executing==1.2.0
fonttools==4.39.3
importlib-metadata==6.3.0
importlib-resources==5.12.0
ipykernel==6.22.0
ipython==8.12.0
jedi==0.18.2
jupyter-client==8.1.0
jupyter-core==5.3.0
kiwisolver==1.4.4
llvmlite==0.39.1
matplotlib==3.7.1
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
numba==0.56.4
numba-progress==0.0.4
numpy==1.23.5
packaging==23.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.2.0
prompt-toolkit==3.0.38
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.15.0
pyparsing==3.0.9
python-dateutil==2.8.2
pyzmq==25.0.2
scipy==1.10.1
six==1.16.0
stack-data==0.6.2
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
typing-extensions==4.5.0
wcwidth==0.2.6
zipp==3.15.0
